T4b/FoundationPose: mesh-controlled vs icpjgr — stronger per-frame pose, weaker default robustness
Track mode: vase 6.6 (rot 3.0!), mug 2.3 (rot 2.7!) DECISIVELY beat icpjgr;
cube tie 18.9; but drifts catastrophically on masher 597/bottle 93. register_each
mode recovers those (bottle 3.2, masher 8.6, spatula 12.8) — all beat icpjgr,
proving FP's per-frame estimator > icpjgr registration; only its default tracker's
robustness lags. Fixed 2 latent fp_track.py bugs (mycpp path shadow, K float32).

+++ render_and_compare/scripts/subprocess_entries/FoundationPose/fp_track.py
@@ -51,7 +51,10 @@ def main():
     logging.getLogger().setLevel(logging.WARNING)
 
     mesh = trimesh.load(a.mesh, process=False)
-    K = np.load(a.K).astype(np.float64)
+    # NOTE: must be float32 -- FoundationPose's compute_crop_window_tf_batch does
+    # `torch.as_tensor(K) @ pts.float()` with no cast, so a float64 K raises
+    # "expected mat1 and mat2 to have the same dtype, but got: double != float".
+    K = np.load(a.K).astype(np.float32)
     frames = sorted(glob.glob(os.path.join(a.frames_dir, "*.jpg")))
     T = len(frames)
